+++ src/components/Navigation/index.tsx
@@ -7,7 +7,7 @@ import { Menu, X } from "lucide-react";
 import { useRef, useState } from "react";
 import { ThemeToggle } from "../theme/themeToggle.tsx";
 import { Button } from "../ui/button.tsx";
-import { NavigationItem } from "./item.tsx";
+import { NavigationItem } from "../navigationItem/index.tsx";
 
 export const Navigation = () => {
 	const [isOpen, setIsOpen] = useState<boolean>(false);
